Fixed code style and other problems
