+++ typos_correction/corrections.txt
@@ -1,0 +1,133 @@
+Identifier           : Correction           : Correction probability
+--------------------------------------------------------------------
+aaaaaaaaaaaaaaa      : aaaaaaaaaaaaaaaa     : 1.000
+enbutton             : nbutton              : 1.000
+stopit               : stopbit              : 1.000
+bxf                  : buf                  : 0.900
+carlos               : carlo                : 0.900
+jazz                 : jazzy                : 0.900
+prepath              : prepatch             : 0.900
+vttablet             : ttablet              : 0.900
+addmul               : addcmul              : 0.800
+gvks                 : gvk                  : 0.800
+lear                 : clear                : 0.800
+logmsg               : longmsg              : 0.800
+mknod                : mknode               : 0.800
+nla                  : nl                   : 0.800
+rebirth              : jrebirth             : 0.800
+reformat             : preformat            : 0.800
+umeric               : numeric              : 0.800
+cproto               : proto                : 0.767
+ighlight             : highlight            : 0.717
+atrix                : matrix               : 0.700
+cryptlib             : cryptolib            : 0.700
+develop              : kdevelop             : 0.700
+fso                  : fs                   : 0.700
+glen                 : len                  : 0.700
+highest              : highesta             : 0.700
+hql                  : sql                  : 0.700
+stress               : tress                : 0.700
+sigtrap              : sigtramp             : 0.675
+iport                : port                 : 0.650
+round                : found                : 0.617
+bunch                : bench                : 0.600
+celsius              : celcius              : 0.600
+dollar               : collar               : 0.600
+effect               : effects              : 0.600
+further              : farther              : 0.600
+goat                 : got                  : 0.600
+hotlist              : hostlist             : 0.600
+lpolestr             : lpcolestr            : 0.600
+mpctx                : pctx                 : 0.600
+sift                 : shift                : 0.600
+subselect            : sbselect             : 0.600
+workbench            : iworkbench           : 0.600
+lots                 : slots                : 0.552
+collie               : collide              : 0.550
+fence                : fences               : 0.540
+protobuf             : jprotobuf            : 0.517
+tmd                  : tmp                  : 0.517
+dbug                 : debug                : 0.500
+openapi              : openapiv             : 0.500
+weakref              : rweakref             : 0.500
+rogram               : program              : 0.458
+getrandom            : geturandom           : 0.445
+second               : seconds              : 0.433
+dasd                 : dash                 : 0.413
+hrtime               : hrtimer              : 0.400
+newbuf               : newbuff              : 0.400
+relinquish           : relenquish           : 0.400
+supr                 : super                : 0.400
+tcheck               : check                : 0.400
+togo                 : gogo                 : 0.400
+umtx                 : mtx                  : 0.400
+unitless             : initless             : 0.400
+wname                : name                 : 0.400
+xpos                 : pos                  : 0.392
+pctl                 : ctl                  : 0.387
+kg                   : g                    : 0.339
+tdb                  : tab                  : 0.333
+alignto              : aligntop             : 0.300
+bukkit               : bukket               : 0.300
+safari               : salari               : 0.300
+subsystem            : subsytem             : 0.300
+mycode               : mcode                : 0.267
+att                  : attr                 : 0.262
+llvm                 : lvm                  : 0.240
+uhalf                : half                 : 0.233
+rw                   : row                  : 0.232
+rtnh                 : rtn                  : 0.217
+definer              : define               : 0.200
+giveref              : xgiveref             : 0.200
+voucher              : toucher              : 0.200
+vulkan               : vulcan               : 0.200
+mdash                : dash                 : 0.197
+sparql               : sparqlq              : 0.195
+xul                  : xml                  : 0.183
+numbers              : number               : 0.170
+dae                  : date                 : 0.169
+red                  : read                 : 0.160
+tpe                  : type                 : 0.160
+bgc                  : gc                   : 0.150
+slick                : slice                : 0.150
+bound                : found                : 0.133
+clu                  : cpu                  : 0.100
+cvttsd               : cvtsd                : 0.100
+dropbox              : cropbox              : 0.100
+endptr               : endpt                : 0.100
+faith                : farith               : 0.100
+lime                 : line                 : 0.100
+matcher              : matched              : 0.100
+pinyin               : pingyin              : 0.100
+pka                  : pkg                  : 0.100
+popcount             : hopcount             : 0.100
+qua                  : lua                  : 0.100
+shield               : sfield               : 0.100
+shows                : show                 : 0.100
+sockaddr             : jsockaddr            : 0.100
+superscript          : esuperscript         : 0.100
+tlib                 : lib                  : 0.100
+verbs                : verts                : 0.100
+xname                : name                 : 0.100
+yydebug              : pydebug              : 0.100
+vha                  : sha                  : 0.070
+condvar              : convar               : 0.067
+econnreset           : connreset            : 0.067
+quantlist            : quantit              : 0.067
+r                    : m                    : 0.057
+avro                 : avio                 : 0.050
+commasep             : comasep              : 0.050
+drag                 : draw                 : 0.050
+enotty               : notty                : 0.050
+gpg                  : gpu                  : 0.050
+rootpath             : bootpath             : 0.050
+trg                  : arg                  : 0.050
+unique               : uniquer              : 0.050
+wild                 : will                 : 0.050
+xhci                 : hci                  : 0.050
+iem                  : item                 : 0.040
+vcard                : card                 : 0.040
+xps                  : ops                  : 0.040
+ptd                  : ptr                  : 0.033
+updater              : update               : 0.022
+ource                : source               : 0.020
