+++ typos_correction/filter_identifiers.py
@@ -1,8 +1,8 @@
 import pandas
 import numpy
- 
 
-def check_split(split, tokens_set): 
+
+def check_split(split, tokens_set):
     if type(split) != str:
         return False
 
@@ -11,20 +11,19 @@ def check_split(split, tokens_set):
             return False
     return True
 
+
 def filter_splitted_identifiers(id_info, tokens_set, out_file):
     """
     Leave rows in a dataframe whose identifiers' tokens are all in tokens_set
     """
     token_split = list(id_info.token_split)
-    filter_array = []
+    filter_array = list(map(lambda x: check_split(x, tokens_set), token_split))
 
-    filter_array = list(map(lambda x:check_split(x, tokens_set), token_split))
-        
     data = id_info.copy()
     data = data[filter_array]
     data.to_pickle(out_file)
-    
-    
+
+
 def filter_identifiers(args):
     id_info = pandas.read_pickle(args.id_file)
     tokens_set = set(numpy.load(args.vocabulary_file))
