+++ typos_correction/get_frequencies.py
@@ -1,6 +1,7 @@
 import numpy
 import pandas
 
+
 def get_frequencies(tokens_set, id_stats, file):
     """
     Dump frequencies of tokens from tokens_set to a file
@@ -9,11 +10,12 @@ def get_frequencies(tokens_set, id_stats, file):
     for token, row in id_stats.iterrows():
         if token in tokens_set:
             frequencies[token] = row.num_occ
-    with open(file, 'w') as f:
+    with open(file, "w") as f:
         for token in list(tokens_set):
-            print(token, requencies[token], file=f)
-            
+            print(token, frequencies[token], file=f)
+
+
 def get_frequencies(args):
-    id_stats = pandas.read_csv(args.stats_file, index_col='identifier')
+    id_stats = pandas.read_csv(args.stats_file, index_col="identifier")
     tokens_set = set(numpy.load(args.vocabulary_file))
     get_frequencies(tokens_set, id_stats, args.out_file)
